+++ CHANGES.rst
@@ -180,6 +180,9 @@ Bug Fixes
 
 - ``astropy.time``
 
+  - Correct UT1->UTC->UT1 round-trip being off by 1 second if UT1 is
+    on a leap second. [#2077]
+
 - ``astropy.units``
 
 - ``astropy.utils``